requirements.txt
ShopifyAPI
pandas
openpyxl
pytest
black
requests
